Add unzipped SunVox DLL contents
NightRadio said it was fine to include the unzipped contents of the zip
as long as the license file was also included. Including this into the
project should make it easier for folks to download the entire
repository and start playing with it. It saves people from the extra
step of downloading the DLL and then unzipping it in the right spot.